requirements.txt
# Core Data Processing
pandas>=2.0.0
openpyxl>=3.1.0

# Pydantic for data validation
pydantic>=2.0.0

# AI/ML APIs
anthropic>=0.7.0
openai>=1.0.0

# Document parsing
PyMuPDF>=1.23.0
python-docx>=1.1.0

# Environment management
python-dotenv>=1.0.0

# Web UI
streamlit>=1.28.0

# Logging and utilities
typing-extensions>=4.0.0